refactor(replication): isolate resync boundary adapters
